Agrego mensajes de error al cargar los productos si no funciona

+++ js/carrito.js
@@ -47,9 +47,16 @@ const eliminarProductoCarrito = () => {
 
 // CARGAR PRODUCTOS DEL CARRITO
 const cargarProductosCarrito = (carritoArrayProductos) => {
-	productosCarrito.innerHTML = "";
-	carritoArrayProductos.forEach((producto) => { productosCarrito.innerHTML += listarProductosCarrito(producto); });
-	eliminarProductoCarrito();
+	if (carritoArrayProductos.length > 0){
+		productosCarrito.innerHTML = "";
+		carritoArrayProductos.forEach((producto) => { productosCarrito.innerHTML += listarProductosCarrito(producto); });
+		eliminarProductoCarrito();
+	} else {
+		productosCarrito.innerHTML = `<div class="errorCarrito">
+										<h3>⛔ Carrito Vacío</h3>
+										<p>Debes Agregar productos al Carrito en Nuestra <a href="../index.html">Home</a></p>
+									</div>`
+	}
 }
 setTimeout(() => {
 	cargarProductosCarrito(carritoArray);
@@ -109,4 +116,9 @@ function vaciarCarrito() {
 		}
 	  })        
 }
-document.querySelector(".vaciarCarrito").addEventListener("click", vaciarCarrito)
+document.querySelector(".vaciarCarrito").addEventListener("click", vaciarCarrito);
+
+// TOOLTIP
+tippy('#myButton', {
+	content: 'motorstore@gmail.com',
+  });
